// src/api/models/GoodOrCamel.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Jaipur
 * API pour le jeu de société Jaipur
 *
 * The version of the OpenAPI document: 1.0
 * Contact: julien@usson.me
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

/**
 * 
 * @export
 * @enum {string}
 */
export enum GoodOrCamel {
    Diamonds = 'diamonds',
    Gold = 'gold',
    Silver = 'silver',
    Cloth = 'cloth',
    Spice = 'spice',
    Leather = 'leather',
    Camel = 'camel'
}

export function GoodOrCamelFromJSON(json: any): GoodOrCamel {
    return GoodOrCamelFromJSONTyped(json, false);
}

export function GoodOrCamelFromJSONTyped(json: any, ignoreDiscriminator: boolean): GoodOrCamel {
    return json as GoodOrCamel;
}

export function GoodOrCamelToJSON(value?: GoodOrCamel | null): any {
    return value as any;
}